requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==23.1.0
    # via gradio
aiohttp==3.8.4
    # via
    #   gradio
    #   openai
aiosignal==1.3.1
    # via aiohttp
altair==5.0.0
    # via gradio
ansiwrap==0.8.4
    # via papermill
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
autopep8==2.0.2
    # via ploomber
backcall==0.2.0
    # via ipython
backoff==2.2.1
    # via posthog
beautifulsoup4==4.12.2
    # via nbconvert
bleach==6.0.0
    # via nbconvert
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   papermill
    #   ploomber
    #   ploomber-core
    #   ploomber-engine
    #   uvicorn
colorama==0.4.6
    # via debuglater
comm==0.1.3
    # via ipykernel
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
debuglater==1.4.4
    # via ploomber-engine
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via papermill
executing==1.2.0
    # via stack-data
fastapi==0.95.2
    # via gradio
fastjsonschema==2.17.1
    # via nbformat
ffmpy==0.3.0
    # via gradio
filelock==3.12.0
    # via huggingface-hub
fonttools==4.39.4
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.5.0
    # via
    #   gradio-client
    #   huggingface-hub
gradio==3.32.0
    # via -r requirements.in
gradio-client==0.2.5
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.2
    # via httpx
httpx==0.24.1
    # via
    #   gradio
    #   gradio-client
huggingface-hub==0.14.1
    # via
    #   gradio
    #   gradio-client
humanize==4.6.0
    # via ploomber
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipdb==0.13.13
    # via ploomber
ipykernel==6.23.1
    # via ploomber
ipython==8.13.2
    # via
    #   ipdb
    #   ipykernel
    #   ploomber
    #   ploomber-engine
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   gradio
    #   nbconvert
    #   ploomber
    #   ploomber-scaffold
jsonschema==4.17.3
    # via
    #   altair
    #   nbformat
jupyter-client==8.2.0
    # via
    #   ipykernel
    #   nbclient
    #   ploomber
jupyter-core==5.3.0
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.2.2
    # via nbconvert
jupytext==1.14.5
    # via ploomber
kiwisolver==1.4.4
    # via matplotlib
linkify-it-py==2.0.2
    # via markdown-it-py
markdown-it-py[linkify]==2.2.0
    # via
    #   gradio
    #   jupytext
    #   mdit-py-plugins
markupsafe==2.1.2
    # via
    #   gradio
    #   jinja2
    #   nbconvert
matplotlib==3.7.1
    # via gradio
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdit-py-plugins==0.3.3
    # via
    #   gradio
    #   jupytext
mdurl==0.1.2
    # via markdown-it-py
mistune==2.0.5
    # via
    #   nbconvert
    #   ploomber
monotonic==1.6
    # via posthog
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nbclient==0.7.4
    # via
    #   nbconvert
    #   papermill
    #   ploomber-engine
nbconvert==7.4.0
    # via ploomber
nbformat==5.8.0
    # via
    #   jupytext
    #   nbclient
    #   nbconvert
    #   papermill
    #   ploomber
    #   ploomber-engine
nest-asyncio==1.5.6
    # via ipykernel
networkx==3.1
    # via ploomber
numpy==1.24.3
    # via
    #   altair
    #   contourpy
    #   gradio
    #   matplotlib
    #   pandas
openai==0.27.7
    # via -r requirements.in
orjson==3.8.13
    # via gradio
packaging==23.1
    # via
    #   gradio-client
    #   huggingface-hub
    #   ipykernel
    #   matplotlib
    #   nbconvert
pandas==2.0.1
    # via
    #   altair
    #   gradio
pandocfilters==1.5.0
    # via nbconvert
papermill==2.4.0
    # via ploomber
parso==0.8.3
    # via
    #   jedi
    #   ploomber
    #   ploomber-engine
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via
    #   gradio
    #   matplotlib
platformdirs==3.5.1
    # via jupyter-core
ploomber==0.22.3
    # via -r requirements.in
ploomber-core==0.2.10
    # via
    #   ploomber
    #   ploomber-engine
ploomber-engine==0.0.28
    # via ploomber
ploomber-scaffold==0.3.1
    # via ploomber
posthog==3.0.1
    # via
    #   ploomber
    #   ploomber-core
prompt-toolkit==3.0.38
    # via ipython
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via autopep8
pydantic==1.10.7
    # via
    #   fastapi
    #   gradio
    #   ploomber
pydub==0.25.1
    # via gradio
pyflakes==3.0.1
    # via ploomber
pygments==2.15.1
    # via
    #   gradio
    #   ipython
    #   nbconvert
    #   ploomber
pyparsing==3.0.9
    # via matplotlib
pypdf==3.9.0
    # via -r requirements.in
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
    #   posthog
python-dotenv==1.0.0
    # via -r requirements.in
python-multipart==0.0.6
    # via gradio
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via
    #   gradio
    #   huggingface-hub
    #   jupytext
    #   papermill
    #   ploomber
    #   ploomber-core
pyzmq==25.0.2
    # via
    #   ipykernel
    #   jupyter-client
regex==2023.5.5
    # via tiktoken
requests==2.30.0
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   papermill
    #   posthog
    #   tiktoken
semantic-version==2.10.0
    # via gradio
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   posthog
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.4.1
    # via beautifulsoup4
sqlalchemy==2.0.15
    # via ploomber
sqlparse==0.4.4
    # via ploomber
stack-data==0.6.2
    # via ipython
starlette==0.27.0
    # via fastapi
tabulate==0.9.0
    # via ploomber
tenacity==8.2.2
    # via papermill
textwrap3==0.9.2
    # via ansiwrap
tiktoken==0.4.0
    # via -r requirements.in
tinycss2==1.2.1
    # via nbconvert
toml==0.10.2
    # via jupytext
tomli==2.0.1
    # via
    #   autopep8
    #   ipdb
toolz==0.12.0
    # via altair
tornado==6.3.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   openai
    #   papermill
    #   ploomber
    #   ploomber-engine
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.5.0
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   sqlalchemy
tzdata==2023.3
    # via pandas
uc-micro-py==1.0.2
    # via linkify-it-py
urllib3==2.0.2
    # via requests
uvicorn==0.22.0
    # via gradio
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websockets==11.0.3
    # via
    #   gradio
    #   gradio-client
yarl==1.9.2
    # via aiohttp
